backend/requirement.txt
fastapi
uvicorn
sqlalchemy
pydantic
pandas
numpy
sentence-transformers
faiss-cpupython-dotenv
requests